@@ -17,7 +17,7 @@
     "start": "webpack serve --open",
     "dev": "webpack serve --open --mode development",
     "test": "react-scripts test --env=jsdom --reporters=default --reporters=jest-junit",
-    "test:coverage": "react-scripts test --coverage --coverageDirectory=build/coverage"
+    "test:coverage": "react-scripts test --coverage --coverageDirectory=build/coverage --watchAll=false"
   },
   "browserslist": [
     ">0.2%",
@@ -63,8 +63,8 @@
     "process": "^0.11.10",
     "prop-types": "^15.7.2",
     "protobufjs": "^6.11.2",
-    "react": "17.0.0",
-    "react-dom": "17.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-i18next": "^11.15.3",
     "react-perfect-scrollbar": "^1.5.8",
@@ -96,10 +96,10 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
-    "jest-fetch-mock": "^3.0.3",
     "jest-junit": "^13.2.0",
     "junit": "^1.4.9",
     "lint-staged": "^12.3.1",
+    "msw": "^0.48.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.5.1",
     "prettier-plugin-packagejson": "^2.2.15",
