@@ -1,12 +1,14 @@
 import React, { Suspense } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.scss';
 // eslint-disable-next-line import/no-unresolved
 import App from './App/App';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <Suspense fallback="...is loading">
     <App />
-  </Suspense>,
-  document.getElementById('root')
+  </Suspense>
 );
